# perf(ext/event): optimize Event constructor