--- a/cli/tests/unit/event_test.ts
+++ b/cli/tests/unit/event_test.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { assert, assertEquals, assertStringIncludes } from "./test_util.ts";
+import { assertEquals, assertStringIncludes } from "./test_util.ts";
 
 Deno.test(function eventInitializedWithType() {
   const type = "click";
@@ -78,19 +78,6 @@ Deno.test(function eventInitializedWithNonStringType() {
   assertEquals(event.type, "undefined");
   assertEquals(event.bubbles, false);
   assertEquals(event.cancelable, false);
-});
-
-// ref https://github.com/web-platform-tests/wpt/blob/master/dom/events/Event-isTrusted.any.js
-Deno.test(function eventIsTrusted() {
-  const desc1 = Object.getOwnPropertyDescriptor(new Event("x"), "isTrusted");
-  assert(desc1);
-  assertEquals(typeof desc1.get, "function");
-
-  const desc2 = Object.getOwnPropertyDescriptor(new Event("x"), "isTrusted");
-  assert(desc2);
-  assertEquals(typeof desc2!.get, "function");
-
-  assertEquals(desc1!.get, desc2!.get);
 });
 
 Deno.test(function eventInspectOutput() {