@marcosc90 2023-08-16
If we want to adhere to the specification, I will remove the changes to `isTrusted`, but this means that the 'Event' constructor will continue to be relatively slow.